renaming out schedulers with the _old tag
