+++ EXOSIMS/SurveySimulation/tieredScheduler_DD_sotoSS.py
@@ -1,4 +1,4 @@
-from EXOSIMS.SurveySimulation.tieredScheduler_SLSQP_old import tieredScheduler_SLSQP_old
+from EXOSIMS.SurveySimulation.tieredScheduler_sotoSS import tieredScheduler_sotoSS
 import EXOSIMS, os
 import astropy.units as u
 import astropy.constants as const
@@ -13,7 +13,7 @@ import time
 import copy
 from EXOSIMS.util.deltaMag import deltaMag
 
-class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
+class tieredScheduler_DD_sotoSS(tieredScheduler_sotoSS):
     """tieredScheduler_DD - tieredScheduler Dual Detection
     
     This class implements a version of the tieredScheduler that performs dual-band
@@ -22,7 +22,7 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
 
     def __init__(self, **specs):
         
-        tieredScheduler_SLSQP_old.__init__(self, **specs)
+        tieredScheduler_sotoSS.__init__(self, **specs)
         
 
     def run_sim(self):
@@ -167,16 +167,6 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
 
                     self.logger.info('  Starshade and telescope aligned at target star')
                     self.vprint('  Starshade and telescope aligned at target star')
-
-                     # PERFORM CHARACTERIZATION and populate spectra list attribute
-                    characterized, char_fZ, char_systemParams, char_SNR, char_intTime = \
-                            self.observation_characterization(sInd, char_mode)
-                    if np.any(characterized):
-                        self.vprint('  Char. results are: %s'%(characterized))
-                    else:
-                        # make sure we don't accidnetally double characterize
-                        TK.advanceToAbsTime(TK.currentTimeAbs.copy() + .01*u.d)
-                    assert char_intTime != 0, "Integration time can't be 0."
                     if np.any(occ_pInds):
                         DRM['char_fEZ'] = SU.fEZ[occ_pInds].to('1/arcsec2').value.tolist()
                         DRM['char_dMag'] = SU.dMag[occ_pInds].tolist()
@@ -184,6 +174,12 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
                     DRM['char_mode'] = dict(char_mode)
                     del DRM['char_mode']['inst'], DRM['char_mode']['syst']
 
+                     # PERFORM CHARACTERIZATION and populate spectra list attribute
+                    characterized, char_fZ, char_systemParams, char_SNR, char_intTime = \
+                            self.observation_characterization(sInd, char_mode)
+                    if np.any(characterized):
+                        self.vprint('  Char. results are: %s'%(characterized))
+                    assert char_intTime != 0, "Integration time can't be 0."
                     # update the occulter wet mass
                     if OS.haveOcculter and char_intTime is not None:
                         DRM = self.update_occulter_mass(DRM, sInd, char_intTime, 'char')
@@ -372,10 +368,8 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
 
             # 2.1 filter out totTimes > integration cutoff
             if len(sInds) > 0:
-                occ_sInds = np.intersect1d(self.occ_intTimeFilterInds, sInds)
-            if len(sInds) > 0:
                 sInds = np.intersect1d(self.intTimeFilterInds, sInds)
-            
+
             # Starttimes based off of slewtime
             occ_startTimes = occ_tmpCurrentTimeAbs.copy() + slewTimes
             occ_startTimesNorm = occ_tmpCurrentTimeNorm.copy() + slewTimes
@@ -386,7 +380,7 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
             # 2.5 Filter stars not observable at startTimes
             try:
                 koTimeInd = np.where(np.round(occ_startTimes[0].value)-self.koTimes.value==0)[0][0]  # find indice where koTime is startTime[0]
-                sInds_occ_ko = occ_sInds[np.where(np.transpose(self.koMap)[koTimeInd].astype(bool)[occ_sInds])[0]]# filters inds by koMap #verified against v1.35
+                sInds_occ_ko = sInds[np.where(np.transpose(self.koMap)[koTimeInd].astype(bool)[sInds])[0]]# filters inds by koMap #verified against v1.35
                 occ_sInds = sInds_occ_ko[np.where(np.in1d(sInds_occ_ko, HIP_sInds))[0]]
             except:#If there are no target stars to observe 
                 sInds_occ_ko = np.asarray([],dtype=int)
@@ -429,10 +423,15 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
                     totTimes = occ_intTimes*char_mode['timeMultiplier']
                     occ_endTimes = occ_startTimes + totTimes
                 else:
-                    occ_intTimes[occ_sInds] = self.calc_targ_intTime(occ_sInds, occ_startTimes[occ_sInds], char_mode)
-                    occ_sInds = occ_sInds[np.where(occ_intTimes[occ_sInds] <= maxIntTime)]  # Filters targets exceeding end of OB
-                    occ_sInds = occ_sInds[np.where(occ_intTimes[occ_sInds] > 0.0*u.d)]  # Filters targets exceeding end of OB
-                    occ_endTimes = occ_startTimes + occ_intTimes
+                    if old_occ_sInd is not None:
+                        occ_sInds, slewTimes[occ_sInds], occ_intTimes[occ_sInds], dV[occ_sInds] = self.refineOcculterSlews(old_occ_sInd, occ_sInds, 
+                                                                                                                       slewTimes, obsTimes, sd, 
+                                                                                                                       char_mode)  
+                        occ_endTimes = tmpCurrentTimeAbs.copy() + occ_intTimes + slewTimes
+                    else:
+                        occ_intTimes[occ_sInds] = self.calc_targ_intTime(occ_sInds, occ_startTimes[occ_sInds], char_mode)
+                        occ_sInds = occ_sInds[np.where(occ_intTimes[occ_sInds] <= maxIntTime)]  # Filters targets exceeding end of OB
+                        occ_endTimes = occ_startTimes + occ_intTimes
                 
                 if maxIntTime.value <= 0:
                     occ_sInds = np.asarray([],dtype=int)
@@ -440,7 +439,6 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
             if len(sInds.tolist()) > 0:
                 intTimes[sInds] = self.calc_targ_intTime(sInds, startTimes[sInds], det_modes[0])
                 sInds = sInds[np.where(intTimes[sInds] <= maxIntTime)]  # Filters targets exceeding end of OB
-                sInds = sInds[np.where(intTimes[sInds] > 0.0*u.d)]
                 endTimes = startTimes + intTimes
                 
                 if maxIntTime.value <= 0:
@@ -475,14 +473,10 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
             sInds = sInds[np.where(np.invert(no_dets))[0]]
 
             # 7 Filter off cornograph stars with too-long inttimes
-            available_time = None
             if self.occ_arrives > TK.currentTimeAbs:
                 available_time = self.occ_arrives - TK.currentTimeAbs.copy()
                 if np.any(sInds[intTimes[sInds] < available_time]):
                     sInds = sInds[intTimes[sInds] < available_time]
-
-            # 8 remove occ targets on ignore_stars list
-            occ_sInds = np.setdiff1d(occ_sInds, self.ignore_stars)
 
             t_det = 0*u.d
             det_mode = copy.deepcopy(det_modes[0])
@@ -524,15 +518,7 @@ class tieredScheduler_DD_SLSQP_old(tieredScheduler_SLSQP_old):
                     det_mode['syst']['optics'] = np.mean((det_mode['syst']['optics'], det_modes[1]['syst']['optics']))
                     det_mode['instName'] = 'combined'
 
-                t_det = self.calc_targ_intTime(np.array([sInd]), np.array([startTimes[sInd]]), det_mode)[0]
-
-                if t_det > maxIntTime and maxIntTime > 0*u.d:
-                    t_det = maxIntTime
-                if available_time is not None and available_time > 0*u.d:
-                    if t_det > available_time:
-                        t_det = available_time.copy().value * u.d
-            else:
-                sInd = None
+                t_det = self.calc_targ_intTime(np.array(sInd), startTimes[sInd], det_mode)[0]
 
             # if no observable target, call the TimeKeeping.wait() method
             if not np.any(sInds) and not np.any(occ_sInds):
