+++ tests/TestSupport/test-scripts/simplest_occ.json
@@ -114,7 +114,7 @@
   ],
   "modules": {
     "PlanetPopulation": " ",
-    "StarCatalog": " ",
+    "StarCatalog": "EXOCAT1",
     "OpticalSystem": " ",
     "ZodiacalLight": " ",
     "BackgroundSources": " ",
